delet unused files and code
