+++ src/types/index.ts
@@ -1,5 +1,5 @@
-// Export all types from their respective files
-export * from './messages';
+// export * from './messages';
 export * from './config';
 export * from './currency';
 export * from './jwt';
+export * from './messagePayload';
